moved and prettified PageHeader
